+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 25.2.0
 - Added GDScript support to the note edit code block syntax highlighter
   (for [#3222](https://github.com/pbek/QOwnNotes/issues/3222), thank you, @DarienMC)
+- Try to fix hypothetically leaking carriage return characters in the headings in
+  the note link dialog under Windows (for [#3221](https://github.com/pbek/QOwnNotes/issues/3221))
 
 ## 25.1.7
 - When inserting an image or attachment file via their insert dialogs or via